feat: tag 함수 분리 및 적용
